move animation spec
