+++ cypress/components/animation/ChainedBalls.svelte
